requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0

# CORS
django-cors-headers==4.3.1

# Image handling
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Database (optional, for PostgreSQL)
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# API Documentation
drf-spectacular==0.26.5

# Pagination
djangorestframework-pagination==0.0.3

# Filtering
django-filter==23.5

# Development
django-extensions==3.2.3